apiext: refactor to make public api clearer
This moves the majority of the code for the apiext server
into the `pkg` folder and hides internal implementation using
the nested `internal` folder to align with standard go conventions
found in most modern projects.

This means only `apiext.Run` and `apiext.PodNamespace` are part
of the public API.

Follow-up commits will do additional refactoring to improve the
structure of the code and its public api.

Signed-off-by: Lance Austin <laustin@datawire.io>